Good riddance, SpecimenUtils
