+++ specimen/src/org/labkey/specimen/actions/SpecimenController.java
@@ -1,5 +1,6 @@
 package org.labkey.specimen.actions;
 
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.json.JSONObject;
@@ -50,12 +51,14 @@ import org.labkey.api.specimen.RequestEventType;
 import org.labkey.api.specimen.RequestedSpecimens;
 import org.labkey.api.specimen.SpecimenManager;
 import org.labkey.api.specimen.SpecimenManagerNew;
+import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenQuerySchema;
 import org.labkey.api.specimen.SpecimenRequestException;
 import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.SpecimenSchema;
 import org.labkey.api.specimen.Vial;
+import org.labkey.api.specimen.actions.IdTypes;
 import org.labkey.api.specimen.actions.ParticipantCommentForm;
 import org.labkey.api.specimen.actions.SpecimenViewTypeForm;
 import org.labkey.api.specimen.importer.RequestabilityManager;
@@ -99,6 +102,7 @@ import org.labkey.api.study.Visit;
 import org.labkey.api.study.model.CohortService;
 import org.labkey.api.study.model.ParticipantDataset;
 import org.labkey.api.study.security.permissions.ManageStudyPermission;
+import org.labkey.api.util.Button;
 import org.labkey.api.util.ConfigurationException;
 import org.labkey.api.util.DateUtil;
 import org.labkey.api.util.ExceptionUtil;
@@ -113,6 +117,7 @@ import org.labkey.api.util.URLHelper;
 import org.labkey.api.util.emailTemplate.EmailTemplateService;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.AjaxCompletion;
+import org.labkey.api.view.DisplayElement;
 import org.labkey.api.view.GridView;
 import org.labkey.api.view.HBox;
 import org.labkey.api.view.HtmlView;
@@ -310,6 +315,201 @@ public class SpecimenController extends SpringActionController
     {
         return (selectedMode == SpecimenQueryView.Mode.COMMENTS) ||
                 (selectedMode == SpecimenQueryView.Mode.DEFAULT && SettingsManager.get().getDisplaySettings(container).isDefaultToCommentsMode());
+    }
+
+    public SpecimenQueryView getSpecimenQueryView(boolean showVials, boolean forExport, @Nullable Collection<? extends ParticipantDataset> cachedFilterData, SpecimenQueryView.Mode viewMode, CohortFilter cohortFilter)
+    {
+        boolean commentsMode = isCommentsMode(getContainer(), viewMode);
+
+        SpecimenQueryView gridView;
+        RepositorySettings settings = SettingsManager.get().getRepositorySettings(getContainer());
+
+        if (cachedFilterData != null)
+        {
+            gridView = SpecimenQueryView.createView(getViewContext(), cachedFilterData,
+                    showVials ? SpecimenQueryView.ViewType.VIALS :
+                            SpecimenQueryView.ViewType.SUMMARY);
+        }
+        else
+        {
+            gridView = SpecimenQueryView.createView(getViewContext(),
+                    showVials ? SpecimenQueryView.ViewType.VIALS :
+                            SpecimenQueryView.ViewType.SUMMARY, cohortFilter);
+        }
+        gridView.setShowHistoryLinks(showVials && !forExport && !settings.isSimple());
+        gridView.setDisableLowVialIndicators(forExport || settings.isSimple());
+        gridView.setShowRecordSelectors(settings.isEnableRequests() || commentsMode); // Need to allow buttons
+        gridView.setShowDeleteButton(false);
+
+        // if we're exporting, we can skip setting up the buttons:
+        if (forExport)
+            return gridView;
+
+        List<DisplayElement> buttons = new ArrayList<>();
+
+        ActionButton ptidListButton = StudyInternalService.get().createParticipantGroupButton(getViewContext(), gridView.getSettings().getDataRegionName(), cohortFilter, false);
+        if (ptidListButton != null)
+            buttons.add(ptidListButton);
+
+        if (settings.isEnableRequests())
+        {
+            MenuButton requestMenuButton = new MenuButton("Request Options");
+            requestMenuButton.addMenuItem("View Existing Requests", SpecimenMigrationService.get().getViewRequestsURL(getContainer()));
+            if (!commentsMode)
+            {
+                if (getViewContext().getContainer().hasPermission(getViewContext().getUser(), RequestSpecimensPermission.class))
+                {
+                    final String jsRegionObject = DataRegion.getJavaScriptObjectReference(gridView.getSettings().getDataRegionName());
+                    String createRequestURL = urlFor(SpecimenMigrationService.get().getShowCreateSpecimenRequestActionClass()).addReturnURL(getViewContext().getActionURL()).toString();
+
+                    requestMenuButton.addMenuItem("Create New Request",
+                            "if (verifySelected(" + jsRegionObject + ".form, '" + createRequestURL +
+                                    "', 'post', 'rows')) " + jsRegionObject + ".form.submit();");
+
+                    if (getViewContext().getContainer().hasPermission(getViewContext().getUser(), ManageRequestsPermission.class) ||
+                            SettingsManager.get().isSpecimenShoppingCartEnabled(getViewContext().getContainer()))
+                    {
+                        requestMenuButton.addMenuItem("Add To Existing Request",
+                                "if (verifySelected(" + jsRegionObject + ".form, '#', " +
+                                        "'get', 'rows')) { " + jsRegionObject + ".getSelected({success: function (data) { showRequestWindow(data.selected, '" + (showVials ? IdTypes.RowId
+                                        : IdTypes.SpecimenHash) + "');}})}");
+                    }
+                }
+            }
+            else
+            {
+                ActionURL endCommentsURL = getViewContext().getActionURL().clone();
+                endCommentsURL.replaceParameter(SpecimenViewTypeForm.PARAMS.viewMode, SpecimenQueryView.Mode.REQUESTS.name());
+                requestMenuButton.addMenuItem("Enable Request Mode", endCommentsURL);
+            }
+
+            buttons.add(requestMenuButton);
+        }
+
+        if (getViewContext().getContainer().hasPermission(getUser(), SetSpecimenCommentsPermission.class))
+        {
+            boolean manualQCEnabled = SettingsManager.get().getDisplaySettings(getViewContext().getContainer()).isEnableManualQCFlagging();
+            if (commentsMode)
+            {
+                MenuButton commentsMenuButton = new MenuButton("Comments" + (manualQCEnabled ? " and QC" : ""));
+                final String jsRegionObject = DataRegion.getJavaScriptObjectReference(gridView.getSettings().getDataRegionName());
+                String setCommentsURL = urlFor(SpecimenMigrationService.get().getUpdateCommentsActionClass()).toString();
+                NavTree setItem = commentsMenuButton.addMenuItem("Set Vial Comment " + (manualQCEnabled ? "or QC State " : "") + "for Selected",
+                        "if (verifySelected(" + jsRegionObject + ".form, '" + setCommentsURL +
+                                "', 'post', 'rows')) " + jsRegionObject + ".form.submit(); return false;");
+                setItem.setId("Comments:Set");
+
+                String clearCommentsURL = urlFor(SpecimenMigrationService.get().getClearCommentsActionClass()).toString();
+                NavTree clearItem = commentsMenuButton.addMenuItem("Clear Vial Comments for Selected",
+                        "if (verifySelected(" + jsRegionObject + ".form, '" + clearCommentsURL +
+                                "', 'post', 'rows') && confirm('This will permanently clear comments for all selected vials. " +
+                                (manualQCEnabled ? "Quality control states will remain unchanged. " : "" )+ "Continue?')) " +
+                                jsRegionObject + ".form.submit();\nreturn false;");
+                clearItem.setId("Comments:Clear");
+
+                ActionURL endCommentsURL = getViewContext().getActionURL().clone();
+                endCommentsURL.replaceParameter(SpecimenViewTypeForm.PARAMS.viewMode, SpecimenQueryView.Mode.REQUESTS.name());
+                NavTree exitItem = commentsMenuButton.addMenuItem("Exit Comments " + (manualQCEnabled ? "and QC " : "") + "mode", endCommentsURL);
+                exitItem.setId("Comments:Exit");
+
+                Study study = getStudy();
+                String subjectNoun = StudyService.get().getSubjectNounSingular(getContainer());
+                boolean addSep = true;
+
+                Integer participantCommentDatasetId = StudyInternalService.get().getParticipantCommentDatasetId(study);
+
+                if (participantCommentDatasetId != null && participantCommentDatasetId != -1)
+                {
+                    Dataset ds = StudyService.get().getDataset(study.getContainer(), participantCommentDatasetId);
+                    if (ds != null && ds.canUpdate(getUser()))
+                    {
+                        if (addSep)
+                        {
+                            commentsMenuButton.addSeparator();
+                            addSep = false;
+                        }
+                        NavTree ptidComments = commentsMenuButton.addMenuItem("Manage " + PageFlowUtil.filter(subjectNoun) + " Comments",
+                            urlProvider(StudyUrls.class).getDatasetURL(getContainer(), participantCommentDatasetId));
+                        ptidComments.setId("Comments:SetParticipant");
+                    }
+                }
+
+                Integer participantVisitCommentDatasetId = StudyInternalService.get().getParticipantVisitCommentDatasetId(study);
+
+                if (participantVisitCommentDatasetId != null && participantVisitCommentDatasetId != -1)
+                {
+                    Dataset ds = StudyService.get().getDataset(study.getContainer(), participantVisitCommentDatasetId);
+                    if (ds != null && ds.canUpdate(getUser()))
+                    {
+                        if (addSep)
+                        {
+                            commentsMenuButton.addSeparator();
+                            addSep = false;
+                        }
+                        NavTree ptidComments = commentsMenuButton.addMenuItem("Manage " + PageFlowUtil.filter(subjectNoun) + "/Visit Comments",
+                            urlProvider(StudyUrls.class).getDatasetURL(getContainer(), participantVisitCommentDatasetId));
+                        ptidComments.setId("Comments:SetParticipantVisit");
+                    }
+                }
+                buttons.add(commentsMenuButton);
+            }
+            else
+            {
+                ActionURL enableCommentsURL = getViewContext().getActionURL().clone();
+                enableCommentsURL.replaceParameter(SpecimenViewTypeForm.PARAMS.viewMode, SpecimenQueryView.Mode.COMMENTS.name());
+                ActionButton commentsButton = new ActionButton("Enable Comments" + (manualQCEnabled ? "/QC" : ""), enableCommentsURL);
+                buttons.add(commentsButton);
+            }
+        }
+
+        if (getViewContext().hasPermission(AdminPermission.class))
+        {
+            Button upload = new Button.ButtonBuilder("Import Specimens")
+                    .href(SpecimenMigrationService.get().getUploadSpecimensURL(getContainer()))
+                    .build();
+            buttons.add(upload);
+        }
+
+        if (settings.isEnableRequests() && getViewContext().getContainer().hasPermission(getViewContext().getUser(), RequestSpecimensPermission.class))
+        {
+            buttons.add(new ButtonBarLineBreak());
+            buttons.add(new ExcludeSiteDropDown());
+        }
+
+        gridView.setButtons(buttons);
+        return gridView;
+    }
+
+    public static final class ExcludeSiteDropDown extends DisplayElement
+    {
+        @Override
+        public void render(RenderContext ctx, Writer out) throws IOException
+        {
+            ActionURL url = ctx.getViewContext().cloneActionURL();
+            url.deleteParameter(SpecimenQueryView.PARAMS.excludeRequestedBySite.name());
+            out.write("Hide Previously Requested By ");
+            out.write("<select ");
+            out.write("onchange=\"window.location=");
+            out.write(PageFlowUtil.jsString(url + "&amp;excludeRequestedBySite="));
+            out.write(" + this.options[this.selectedIndex].value;\"");
+            out.write(">");
+            out.write("<option value=''>&lt;Show All&gt;</option>");
+            String excludeStr = ctx.getRequest().getParameter(SpecimenQueryView.PARAMS.excludeRequestedBySite.name());
+            int locationId = null == StringUtils.trimToNull(excludeStr) ? 0 : Integer.parseInt(excludeStr);
+            List<LocationImpl> locations = LocationManager.get().getValidRequestingLocations(ctx.getContainer());
+            for (LocationImpl location : locations)
+            {
+                out.write("<option value=\"");
+                out.write(String.valueOf(location.getRowId()));
+                out.write("\"");
+                if (location.getRowId() == locationId)
+                    out.write(" SELECTED ");
+                out.write("\">");
+                out.write(PageFlowUtil.filter(location.getDisplayName()));
+                out.write("</option>");
+            }
+            out.write("</select>");
+        }
     }
 
     public GridView getRequestEventGridView(HttpServletRequest request, BindException errors, SimpleFilter filter)
@@ -4140,7 +4340,7 @@ public class SpecimenController extends SpringActionController
 
             _vialView = form.isShowVials();
             CohortFilter cohortFilter = CohortService.get().getFromURL(getContainer(), getUser(), getViewContext().getActionURL(), _vialView ? "SpecimenDetail" : "SpecimenSummary");
-            SpecimenQueryView view = StudyInternalService.get().getSpecimenQueryView(getViewContext(), _vialView, forExport, null, form.getViewModeEnum(), cohortFilter);
+            SpecimenQueryView view = getSpecimenQueryView(_vialView, forExport, null, form.getViewModeEnum(), cohortFilter);
             if (isCommentsMode(getContainer(), form.getViewModeEnum()))
                 view.setRestrictRecordSelectors(false);
             return view;
@@ -5380,10 +5580,10 @@ public class SpecimenController extends SpringActionController
             CohortFilter cohortFilter = CohortService.get().getFromURL(getContainer(), getUser(), getViewContext().getActionURL(), SpecimenQueryView.ViewType.SUMMARY.getQueryName());
             if (lsids != null)
             {
-                view = StudyInternalService.get().getSpecimenQueryView(getViewContext(), form.isShowVials(), forExport, getFilterPds(), form.getViewModeEnum(), cohortFilter);
+                view = getSpecimenQueryView(form.isShowVials(), forExport, getFilterPds(), form.getViewModeEnum(), cohortFilter);
             }
             else
-                view = StudyInternalService.get().getSpecimenQueryView(getViewContext(), form.isShowVials(), forExport, null, form.getViewModeEnum(), cohortFilter);
+                view = getSpecimenQueryView(form.isShowVials(), forExport, null, form.getViewModeEnum(), cohortFilter);
             view.setAllowExportExternalQuery(false);
             return view;
         }
